refactor(dashboard/ide): drop legacy -mock suffix on real-backend panels
ide-conversation-rail and ide-interject components were wired to real
backend (fetch /api/v1/board, fetchKeeperDecisions, fetchCascadeStrategyTrace,
interject-store) but their filenames and class names still carried the
-mock / Mock suffix from the original mockup stage. This caused readers
to assume the dashboard right rail was displaying fake data.

Pure rename + class symbol update + import path update; no behavior change.

- 6 files renamed (src + .test + .a11y.test for each of 2 components)
- IdeConversationRailMock -> IdeConversationRail
- IdeInterjectMock -> IdeInterject
- 3 trace-bridge JSDoc comment references updated
- ide-shell.ts / ide-shell.test.ts import + usage updated
- design-system/CHANGELOG + RFC-0022 path references updated
- ide-activity-mock was already renamed to ide-activity-panel on main

tsc --noEmit: clean for changed paths (one pre-existing unrelated error
in config/navigation.ts present on origin/main).
vitest: 5/5 affected suites pass (36/36 tests).
